Move sourceignore its own dir
