+++ sourceignore/go.mod
@@ -1,0 +1,19 @@
+module github.com/fluxcd/pkg/sourceignore
+
+go 1.18
+
+require (
+	github.com/go-git/go-git/v5 v5.4.2
+	gotest.tools v2.2.0+incompatible
+)
+
+require (
+	github.com/acomagu/bufpipe v1.0.3 // indirect
+	github.com/go-git/gcfg v1.5.0 // indirect
+	github.com/go-git/go-billy/v5 v5.3.1 // indirect
+	github.com/google/go-cmp v0.3.0 // indirect
+	github.com/jbenet/go-context v0.0.0-20150711004518-d14ea06fba99 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
+	golang.org/x/net v0.0.0-20210326060303-6b1517762897 // indirect
+	gopkg.in/warnings.v0 v0.1.2 // indirect
+)
